Switch to using doubles for all channel/PID/motor variables

+++ src/flight/Mixer.h
@@ -11,5 +11,5 @@ namespace Mixer {
      * Requires configuration to be loaded.
      */
     void initialize();
-    void applyMix(int16_t throttle, double pitchPidOutput, double rollPidOutput, double yawPidOutput);
+    void applyMix(double throttle, double pitchPidOutput, double rollPidOutput, double yawPidOutput);
 }
