+++ src/hardware/drivers/motor/DShotMotor.h
@@ -6,6 +6,6 @@
 namespace MotorDrivers {
     class DShotMotor : public Hardware::Motor {
     public:
-        void setOutput(int16_t output) override;
+        void setOutput(double output) override;
     };
 }
